Split `list_unchop()` from `vec_chop()`
So we can `@keyword internal` it and isolate it more
